Modernize Code, Enhance performance & Improve Debug API
- Refactored codebase using modern C++ best practices.
- Optimized resource allocation in debug event processing.
- Enhanced Debug API for better stability and maintainability.

+++ src/include/timeout.h
@@ -87,7 +87,7 @@ VOID CALLBACK TimerRoutine(PVOID lpParameter, BOOLEAN TimerOrWaitFired) {
 
 VOID WaitForInputOrTimeout(
     HANDLE     hStdout,
-    LONG       dwTimeout,
+    DWORD       dwTimeout,
     BOOL       bConsole
 ) {
     HANDLE hStdin = RtlStandardInput();
